refactor: renamed utils into test_util
