+++ meson.build
@@ -124,13 +124,13 @@ alltests = executable('alltests',
 	'test/webfuse/tests/core/jsonrpc/test_response_parser.cc',
 	'test/webfuse/tests/core/timer/test_timepoint.cc',
 	'test/webfuse/tests/core/timer/test_timer.cc',
-	'test/webfuse/utils/tempdir.cc',
-	'test/webfuse/utils/file_utils.cc',
-	'test/webfuse/utils/timeout_watcher.cc',
-	'test/webfuse/utils/ws_server.cc',
-	'test/webfuse/utils/ws_server2.cc',
-	'test/webfuse/utils/adapter_client.cc',
-	'test/webfuse/utils/jansson_test_environment.cc',
+	'test/webfuse/test_util/tempdir.cc',
+	'test/webfuse/test_util/file_utils.cc',
+	'test/webfuse/test_util/timeout_watcher.cc',
+	'test/webfuse/test_util/ws_server.cc',
+	'test/webfuse/test_util/ws_server2.cc',
+	'test/webfuse/test_util/adapter_client.cc',
+	'test/webfuse/test_util/jansson_test_environment.cc',
 	'test/webfuse/mocks/mock_authenticator.cc',
 	'test/webfuse/mocks/mock_fuse.cc',
 	'test/webfuse/mocks/mock_operation_context.cc',
