+++ test/webfuse/test_util/tempdir.cc
@@ -1,4 +1,4 @@
-#include "webfuse/utils/tempdir.hpp"
+#include "webfuse/test_util/tempdir.hpp"
 
 #include <unistd.h>
 #include <cstdio>
