+++ test/webfuse/test_util/timeout_watcher.cc
@@ -1,4 +1,4 @@
-#include "webfuse/utils/timeout_watcher.hpp"
+#include "webfuse/test_util/timeout_watcher.hpp"
 #include <stdexcept>
 #include <thread>
