+++ test/webfuse/test_util/ws_server.h
@@ -1,5 +1,5 @@
-#ifndef WF_TEST_UTILS_WS_SERVER_HPP
-#define WF_TEST_UTILS_WS_SERVER_HPP
+#ifndef WF_TEST_UTIL_WS_SERVER_HPP
+#define WF_TEST_UTIL_WS_SERVER_HPP
 
 #include <jansson.h>
 #include <string>
