+++ test/webfuse/test_util/ws_server2.hpp
@@ -1,5 +1,5 @@
-#ifndef WF_TEST_UTILS_WS_SERVER2_HPP
-#define WF_TEST_UTILS_WS_SERVER2_HPP
+#ifndef WF_TEST_UTIL_WS_SERVER2_HPP
+#define WF_TEST_UTIL_WS_SERVER2_HPP
 
 #include <jansson.h>
 #include <string>
