+++ test/webfuse/tests/core/jsonrpc/mock_timer.cc
@@ -1,5 +1,5 @@
 #include "webfuse/tests/core/jsonrpc/mock_timer.hpp"
-#include "webfuse/utils/wrap.hpp"
+#include "webfuse/test_util/wrap.hpp"
 
 extern "C"
 {
